Tracker Working
Currently, info_hash is just the direct info_hash from the torrent file without any SHA1.

I'm having difficult figuring out how to have the peer extract the DHT table from the tracker at the right time so that the nodes are updated before the downloads start.
Maybe this needs to be threaded as well, but I'm unsure.